pyproject.toml
[project]
name = "frenchtrainer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.3",
    "chainlit>=2.2.1",
    "edge-tts>=7.0.0",
    "fastapi[standard]>=0.115.8",
    "jinja2>=3.1.5",
    "marimo>=0.11.8",
    "marisa-trie>=1.2.1",
    "matplotlib>=3.10.0",
    "ollama>=0.4.7",
    "pandas>=2.2.3",
    "playsound3>=2.5.2",
    "python-decouple>=3.8",
    "ruff>=0.9.7",
    "seaborn>=0.13.2",
    "sqlite-utils>=3.38",
    "sqlmodel>=0.0.22",
]
